API/requirements.txt
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-speech==2.13.1
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
python-decouple==3.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
